multiple sources

+++ Ray_tracer.h
@@ -9,7 +9,7 @@ class Camera
 public :
     Vector_t position;
     Vector_t direction;
-    int FOV;
+    float FOV;
 
 public :
     Camera(Vector_t pos, Vector_t dir, int fov) :
